Superficial changes and debug config.yaml
Now using fixed meteorology and first-order integration for the points to allow faster testing. Complex contrails enabled too. Alongside this we have a host of visual improvements, specifically the implementation of a separate queue in the idle simulator for user-created points (fixed lifetime - looks nicer) and the hiding of points until they are updated (avoids blinking as they were visible for a single frame).